requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
pypdf